requirements.txt
streamlit==1.32.0
google-generativeai==0.5.0
anthropic==0.21.3
python-dotenv==1.0.1
